+++ src/main/java/com/projects/supporthub/service/implementation/UserServiceImpl.java
@@ -4,6 +4,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import com.projects.supporthub.exception.UserNotFoundException;
 import com.projects.supporthub.model.User;
 import com.projects.supporthub.repository.TicketRepository;
 import com.projects.supporthub.repository.UserRepository;
@@ -28,12 +29,23 @@ public class UserServiceImpl implements UserService
 
     public void deleteUserById(String id)
     {
-        userRepo.deleteById(id);
-        ticketRepo.deleteByCreatorId(id);
+        if (userRepo.findById(id).isEmpty())
+        {
+            throw new UserNotFoundException("User not found.");
+        }
+        else
+        {
+            userRepo.deleteById(id);
+            ticketRepo.deleteByCreatorId(id);
+        }
     }
 
     public User getUserById(String id)
     {
+        if (userRepo.findById(id).isEmpty())
+        {
+            throw new UserNotFoundException("User not found.");
+        }
         return userRepo.findById(id).get();
     }
